Install Django REST Framework
Create a text file called requirements which lists the Python packages in the environment, add Django Filters and Rest Framework to the INSTALLED_APPS item in the project's settings.py file and create a configuration dictionary for the REST framework API's global settings.

As for the global settings themselves, the default authentication will use Session Authentication, and the default permission will use the IsAuthenticatedOrReadOnly policy.

For the user to be abble to login and logout using the REST framework authentication views, point the root URLconf at the rest_framework.urls.

These changes are needed for setting up the project for bulding the API.