Customize generic views
Overwrite AvaliacoesAPIView to get and return to the user all ratings of an specified course if the course's id was given, otherwise, if no id was given, return to the user all of the ratings stored in the database. Similarly, overwrite AvaliacaoAPIView to get an specific rating of an specified course if any rating's id was given, otherwise, if no id was given, return to the user all ratings of an specified course. Finally, alter the cursos/urls.py file for the app's urls to cover all these new possibilities of requests.

Needed for applying complex business rule which generic views were not able to fully meet.